hotfix: new structure of folders
a little bit less extra folders.
