+++ additional_tests/clear_set/clear_set.pdp
