+++ include/pdp_add_func.h
@@ -1,0 +1,8 @@
+#pragma once
+
+// Функция логирования
+void logger(int log_lvl, const char * format, ...);
+// Печать всех регистров
+void reg_print();
+void psw_log_print();
+void result_print();
