--- pdp-11/pr/include/pdp_tests.h
@@ -1,7 +1,0 @@
-#pragma once
-
-void test_mem();
-void test_odd_adr();
-void test_logger_prints();
-void test_logger();
-void test_console_arg();
